Team17/requirements.txt
flask
flask-sqlalchemy
python-dotenv
groq
werkzeug